requirements (1).txt
flask==3.1.0
google-generativeai==0.8.3
gunicorn==23.0.0
requests==2.32.3
